python-services/summary-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
langchain==0.3.27
langchain-google-genai==2.1.9
